requirements.txt
Spark 
Python
Pandas